chore: fixed value, make error messages more friendly and put all tiny classes in one
