--- kngsild/src/main/kotlin/io/egm/kngsild/model/AuthType.kt
@@ -1,7 +1,0 @@
-package io.egm.kngsild.model
-
-enum class AuthType {
-    NONE,
-    PROVIDED_TOKEN,
-    CLIENT_CREDENTIALS,
-}
