--- kngsild/src/main/kotlin/io/egm/kngsild/model/ClientCredentials.kt
@@ -1,8 +1,0 @@
-package io.egm.kngsild.model
-
-data class ClientCredentials(
-    val serverUrl: String,
-    val clientId: String,
-    val clientSecret: String,
-    val grantType: String
-)
