--- kngsild/src/main/kotlin/io/egm/kngsild/model/ProvidedToken.kt
@@ -1,3 +1,0 @@
-package io.egm.kngsild.model
-
-data class ProvidedToken(val accessToken: String)
